Make CSSThemeManager really only focus on theme
The settings repo now handles caching itself. Since it is the only IO to settings records, it should work well without issue.

This also reduces the scope of the css theme manager to only deal with handling theme data. Realistically, we can also remove the css theme manager but let's leave that for another PR